+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - delete a volume: add assigned papers error's message.
 - Sorted paper by volume: include assigned papers to this one as a secondary volume
-- [#445](https://github.com/CCSDForge/episciences/issues/445): do not limit the number of papers per page to 10: the default limit (100) is used. 
+- [#445](https://github.com/CCSDForge/episciences/issues/445): do not limit the number of papers per page to 10: the default limit (100) is used + acceptance date descending order. 
 ### Fixed
 - [#423](https://github.com/CCSDForge/episciences/issues/423): fix typo Technical support Email. 
 - [#443](https://github.com/CCSDForge/episciences/issues/443): In the templates (New version - author and co-author copy, New temporary version author and co-author copy) the tag %%PERMANENT_ARTICLE_ID%% is used, but it did not appear in the subject of the message.